requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
bandit==1.8.3
bcrypt==4.0.1
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
click==8.2.0
cryptography==45.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
git-filter-repo==2.47.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pbr==6.1.1
platformdirs==4.3.8
pluggy==1.6.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.3.5
pytest-asyncio==1.0.0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rsa==4.9.1
ruff==0.11.11
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
stevedore==5.4.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
